# openapi.yaml
components:
  schemas:
...
    Analysis:
      description: An Analysis is a data transformation that transforms input data
        to output data. The workflow used to achieve this transformation and the individual
        steps are also captured.
      properties:
        accession:
          description: A unique GHGA identifier assigned to an entity for the sole
            purpose of referring to that entity in a global scope.
          title: Accession
          type: string
        alias:
          description: An alias uniquely identifying this Analysis entitiy.
          title: Alias
          type: string
        creation_date:
          description: Timestamp (in ISO 8601 format) when the entity was created.
          title: Creation Date
          type: string
        description:
          description: 'Describing how an Analysis was carried out. (e.g.: computational
            tools, settings, etc.).'
          title: Description
          type: string
        ega_accession:
          description: A unique European Genome-Phenome Archive (EGA) identifier assigned
            to an entity for the sole purpose of referring to that entity within the
            EGA federated network.
          title: Ega Accession
          type: string
        has_analysis_process:
          anyOf:
          - items:
              $ref: '#/components/schemas/AnalysisProcess'
            type: array
          - items:
              type: string
            type: array
          description: One or more Analysis Process entities associated with this
            Analysis.
          title: Has Analysis Process
        has_input:
          anyOf:
          - items:
              $ref: '#/components/schemas/File'
            type: array
          - items:
              type: string
            type: array
          description: The input data File entities used in the Analysis.
          title: Has Input
        has_output:
          anyOf:
          - items:
              $ref: '#/components/schemas/File'
            type: array
          - items:
              type: string
            type: array
          description: The output data File entities generated by this Analysis.
          title: Has Output
        has_study:
          anyOf:
          - $ref: '#/components/schemas/Study'
          - type: string
          description: The Study entity associated with this Analysis.
          title: Has Study
        has_workflow:
          anyOf:
          - items:
              $ref: '#/components/schemas/Workflow'
            type: array
          - items:
              type: string
            type: array
          description: One or more Workflow entities associated with this Analysis.
          title: Has Workflow
        id:
          description: An identifier that uniquely represents an entity.
          title: Id
          type: string
        reference_chromosome:
          description: The reference chromosome used for this Analysis.
          title: Reference Chromosome
          type: string
        reference_genome:
          description: 'A published genetic sequence that is used as a reference sequence
            against which other sequences are compared. Reference genome(s) or annotation(s)
            used for prior analyses (eg: GRCh38.p13).'
          title: Reference Genome
          type: string
        schema_type:
          enum:
          - Analysis
          title: Schema Type
          type: string
        schema_version:
          description: The version of the schema an instance corresponds to.
          title: Schema Version
          type: string
        title:
          description: The title that describes an entity.
          title: Title
          type: string
        type:
          description: The type of the Analysis. Either Reference Alignment (BAM)
            or Sequence Variation (VCF)
          title: Type
          type: string
        update_date:
          description: Timestamp (in ISO 8601 format) when the entity was updated.
          title: Update Date
          type: string
        xref:
          description: Database cross references for an entity.
          items:
            type: string
          title: Xref
          type: array
      required:
      - schema_type
      title: Analysis
      type: object
    AnalysisProcess:
      description: An analysis process is a process that describes how one or more
        Files, from a Study, are transformed to another set of Files via a Workflow.
        The analysis process also keeps track of the workflow metadata and the Agent
        that is running the Analysis.
      properties:
        alias:
          description: The alias for an entity.
          title: Alias
          type: string
        creation_date:
          description: Timestamp (in ISO 8601 format) when the entity was created.
          title: Creation Date
          type: string
        has_agent:
          anyOf:
          - $ref: '#/components/schemas/Agent'
          - type: string
          description: The Agent - a software, institution, or human - that is executing
            or responsible for executing the workflow.
          title: Has Agent
        has_input:
          anyOf:
          - items:
              $ref: '#/components/schemas/File'
            type: array
          - items:
              type: string
            type: array
          description: The input data File entities used in the Analysis Process.
          title: Has Input
        has_output:
          anyOf:
          - items:
              $ref: '#/components/schemas/File'
            type: array
          - items:
              type: string
            type: array
          description: The output data File entities generated by the Analysis Process.
          title: Has Output
        has_workflow:
          anyOf:
          - $ref: '#/components/schemas/Workflow'
          - type: string
          description: The Workflow entity associated with this Analysis Process.
          title: Has Workflow
        id:
          description: An identifier that uniquely represents an entity.
          title: Id
          type: string
        schema_type:
          enum:
          - AnalysisProcess
          title: Schema Type
          type: string
        schema_version:
          description: The version of the schema an instance corresponds to.
          title: Schema Version
          type: string
        title:
          description: The title that describes an entity.
          title: Title
          type: string
        update_date:
          description: Timestamp (in ISO 8601 format) when the entity was updated.
          title: Update Date
          type: string
        xref:
          description: Database cross references for an entity.
          items:
            type: string
          title: Xref
          type: array
      required:
      - schema_type
      title: AnalysisProcess
      type: object
...
    Biospecimen:
      description: A Biospecimen is any natural material taken from a biological entity
        (usually a human) for testing, diagnostics, treatment, or research purposes.
        The Biospecimen is linked to the Individual from which the Biospecimen is
        derived.
      properties:
        accession:
          description: A unique GHGA identifier assigned to an entity for the sole
            purpose of referring to that entity in a global scope.
          title: Accession
          type: string
        alias:
          description: The alias for an entity.
          title: Alias
          type: string
        creation_date:
          description: Timestamp (in ISO 8601 format) when the entity was created.
          title: Creation Date
          type: string
        description:
          description: Description of an entity.
          title: Description
          type: string
        has_individual:
          anyOf:
          - oneOf:
            - $ref: '#/components/schemas/Donor'
            - $ref: '#/components/schemas/Individual'
          - type: string
          description: The Individual entity from which this Biospecimen was derived.
          title: Has Individual
        id:
          description: An identifier that uniquely represents an entity.
          title: Id
          type: string
        isolation:
          description: Method or device employed for collecting/isolating a biospecimen
            or a sample.
          title: Isolation
          type: string
        name:
          description: The name for an entity.
          title: Name
          type: string
        schema_type:
          enum:
          - Biospecimen
          title: Schema Type
          type: string
        schema_version:
          description: The version of the schema an instance corresponds to.
          title: Schema Version
          type: string
        storage:
          description: Methods by which a biospecimen or a sample is stored (e.g.
            frozen in liquid nitrogen).
          title: Storage
          type: string
        type:
          description: The type of Biospecimen.
          title: Type
          type: string
        update_date:
          description: Timestamp (in ISO 8601 format) when the entity was updated.
          title: Update Date
          type: string
        xref:
          description: Database cross references for an entity.
          items:
            type: string
          title: Xref
          type: array
      required:
      - schema_type
      title: Biospecimen
      type: object
...
    CreateDataAccessCommittee:
      description: A group of members that are delegated to grant access to one or
        more datasets after ensuring the minimum criteria for data sharing has been
        met, and request for data use does not raise ethical and/or legal concerns.
      properties:
        accession:
          description: A unique GHGA identifier assigned to an entity for the sole
            purpose of referring to that entity in a global scope.
          title: Accession
          type: string
        alias:
          description: The alias for an entity.
          title: Alias
          type: string
        description:
          description: A description for the Data Access Committee.
          title: Description
          type: string
        ega_accession:
          description: A unique European Genome-Phenome Archive (EGA) identifier assigned
            to an entity for the sole purpose of referring to that entity within the
            EGA federated network.
          title: Ega Accession
          type: string
        has_attribute:
          description: Key/value pairs corresponding to an entity.
          items:
            $ref: '#/components/schemas/metadata_repository_service__creation_models__Attribute'
          title: Has Attribute
          type: array
        has_member:
          anyOf:
          - items:
              $ref: '#/components/schemas/CreateMember'
            type: array
          - items:
              type: string
            type: array
          description: All the members that are part of this Data Access Committee.
          title: Has Member
        main_contact:
          anyOf:
          - $ref: '#/components/schemas/CreateMember'
          - type: string
          description: The main contact for the Data Access Committee.
          title: Main Contact
        name:
          description: The name for the Data Access Committee.
          title: Name
          type: string
        schema_type:
          enum:
          - CreateDataAccessCommittee
          title: Schema Type
          type: string
        schema_version:
          description: The version of the schema an instance corresponds to.
          title: Schema Version
          type: string
        xref:
          description: Database cross references for an entity.
          items:
            type: string
          title: Xref
          type: array
      required:
      - schema_type
      title: CreateDataAccessCommittee
      type: object
...
    DataAccessCommittee:
      description: A group of members that are delegated to grant access to one or
        more datasets after ensuring the minimum criteria for data sharing has been
        met, and request for data use does not raise ethical and/or legal concerns.
      properties:
        accession:
          description: A unique GHGA identifier assigned to an entity for the sole
            purpose of referring to that entity in a global scope.
          title: Accession
          type: string
        alias:
          description: The alias for an entity.
          title: Alias
          type: string
        creation_date:
          description: Timestamp (in ISO 8601 format) when the entity was created.
          title: Creation Date
          type: string
        description:
          description: A description for the Data Access Committee.
          title: Description
          type: string
        ega_accession:
          description: A unique European Genome-Phenome Archive (EGA) identifier assigned
            to an entity for the sole purpose of referring to that entity within the
            EGA federated network.
          title: Ega Accession
          type: string
        has_attribute:
          description: Key/value pairs corresponding to an entity.
          items:
            $ref: '#/components/schemas/metadata_repository_service__models__Attribute'
          title: Has Attribute
          type: array
        has_member:
          anyOf:
          - items:
              $ref: '#/components/schemas/Member'
            type: array
          - items:
              type: string
            type: array
          description: All the members that are part of this Data Access Committee.
          title: Has Member
        id:
          description: An identifier that uniquely represents an entity.
          title: Id
          type: string
        main_contact:
          anyOf:
          - $ref: '#/components/schemas/Member'
          - type: string
          description: The main contact for the Data Access Committee.
          title: Main Contact
        name:
          description: The name for the Data Access Committee.
          title: Name
          type: string
        schema_type:
          enum:
          - DataAccessCommittee
          title: Schema Type
          type: string
        schema_version:
          description: The version of the schema an instance corresponds to.
          title: Schema Version
          type: string
        update_date:
          description: Timestamp (in ISO 8601 format) when the entity was updated.
          title: Update Date
          type: string
        xref:
          description: Database cross references for an entity.
          items:
            type: string
          title: Xref
          type: array
      required:
      - schema_type
      title: DataAccessCommittee
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
    Workflow:
      description: A Workflow is an abstraction that represents the workflow used
        to perform an analysis. The Workflow entity captures workflow-specific attributes
        that are relevant for an Analysis entity. The Workflow entity may be further
        characterized by its children where each child has fields that are relevant
        to that particular workflow.
      properties:
        alias:
          description: The alias for an entity.
          title: Alias
          type: string
        creation_date:
          description: Timestamp (in ISO 8601 format) when the entity was created.
          title: Creation Date
          type: string
        has_workflow_step:
          anyOf:
          - items:
              $ref: '#/components/schemas/WorkflowStep'
            type: array
          - items:
              type: string
            type: array
          description: The individual workflow step that with other workflow step(s)
            collectively defines a Workflow entity.
          title: Has Workflow Step
        id:
          description: An identifier that uniquely represents an entity.
          title: Id
          type: string
        name:
          description: The name for an entity.
          title: Name
          type: string
        schema_type:
          enum:
          - Workflow
          title: Schema Type
          type: string
        schema_version:
          description: The version of the schema an instance corresponds to.
          title: Schema Version
          type: string
        update_date:
          description: Timestamp (in ISO 8601 format) when the entity was updated.
          title: Update Date
          type: string
        xref:
          description: Database cross references for an entity.
          items:
            type: string
          title: Xref
          type: array
      required:
      - schema_type
      title: Workflow
      type: object
...
info:
  title: FastAPI
  version: 0.1.0
openapi: 3.0.2
paths:
...
  /analyses/{analysis_id}:
    get:
      description: Given an Analysis ID, get the Analysis record from the metadata
        store.
      operationId: get_analyses_analyses__analysis_id__get
      parameters:
      - in: path
        name: analysis_id
        required: true
        schema:
          title: Analysis Id
          type: string
      - in: query
        name: embedded
        required: false
        schema:
          default: false
          title: Embedded
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Analysis'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get an Analysis
      tags:
      - Query
  /analysis_process/{analysis_process_id}:
    get:
      description: Given an AnalysisProcess ID, get the AnalysisProcess record from
        the metadata store.
      operationId: get_analysis_processes_analysis_process__analysis_process_id__get
      parameters:
      - in: path
        name: analysis_process_id
        required: true
        schema:
          title: Analysis Process Id
          type: string
      - in: query
        name: embedded
        required: false
        schema:
          default: false
          title: Embedded
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisProcess'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get an AnalysisProcess
      tags:
      - Query
  /biospecimens/{biospecimen_id}:
    get:
      description: Given a Biospecimen ID, get the Biospecimen record from the metadata
        store.
      operationId: get_biospecimens_biospecimens__biospecimen_id__get
      parameters:
      - in: path
        name: biospecimen_id
        required: true
        schema:
          title: Biospecimen Id
          type: string
      - in: query
        name: embedded
        required: false
        schema:
          default: false
          title: Embedded
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Biospecimen'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get a Biospecimen
      tags:
      - Query
  /data_access_committees:
    post:
      description: Create a DataAccessCommittee and write to the metadata store.
      operationId: create_data_access_committees_data_access_committees_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDataAccessCommittee'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataAccessCommittee'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create a DataAccessCommittee
      tags:
      - DataAccessCommittee
  /data_access_committees/{data_access_committee_id}:
    get:
      description: 'Given a DataAccessCommittee ID, get the DataAccessCommittee record

        from the metadata store.'
      operationId: get_data_access_committees_data_access_committees__data_access_committee_id__get
      parameters:
      - in: path
        name: data_access_committee_id
        required: true
        schema:
          title: Data Access Committee Id
          type: string
      - in: query
        name: embedded
        required: false
        schema:
          default: false
          title: Embedded
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataAccessCommittee'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get a DataAccessCommittee
      tags:
      - Query
  /data_access_policies:
    post:
      description: Create a DataAccessPolicy and write to the metadata store.
      operationId: create_data_access_policies_data_access_policies_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDataAccessPolicy'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataAccessPolicy'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create a DataAccessPolicy
      tags:
      - DataAccessPolicy
  /data_access_policies/{data_access_policy_id}:
    get:
      description: Given a DataAccessPolicy ID, get the DataAccessPolicy record from
        the metadata store.
      operationId: get_data_access_policies_data_access_policies__data_access_policy_id__get
      parameters:
      - in: path
        name: data_access_policy_id
        required: true
        schema:
          title: Data Access Policy Id
          type: string
      - in: query
        name: embedded
        required: false
        schema:
          default: false
          title: Embedded
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataAccessPolicy'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get a DataAccessPolicy
      tags:
      - Query
...